--- src/components/Login.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Template from "./Template";
-
-function Login () {
-    return (
-        <div>
-            <Template 
-            formtype='login' 
-            />
-        </div>
-    );
-}
-
-export default Login;
