--- src/components/Parentpage.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { Outlet } from "react-router-dom";
-
-function Parentpage () {
-    return (
-        <div>
-            <Outlet />
-        </div>
-    );
-}
-
-export default Parentpage;
