--- src/components/Signup.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Template from "./Template";
-
-function Signup () {
-    return (
-        <div>
-            <Template 
-            formtype='signup'
-            />
-        </div>
-    );
-}
-
-export default Signup;
