+++ src/components/pages/AboutUs.js
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { AppContext } from "./context/AppContext";
+import { AppContext } from "../context/AppContext";
 
 function AboutUs() {
   const {sidebar, setSidebar} = useContext(AppContext);
