+++ src/components/pages/Login.js
@@ -1,0 +1,14 @@
+import React from "react";
+import Template from "./Template";
+
+function Login () {
+    return (
+        <div>
+            <Template 
+            formtype='login' 
+            />
+        </div>
+    );
+}
+
+export default Login;
