+++ src/components/pages/Parentpage.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { Outlet } from "react-router-dom";
+
+function Parentpage () {
+    return (
+        <div>
+            <Outlet />
+        </div>
+    );
+}
+
+export default Parentpage;
