+++ src/components/pages/Sidebar.js
@@ -1,6 +1,6 @@
 import React , { useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import { AppContext } from "./context/AppContext";
+import { AppContext } from "../context/AppContext";
 
 function Sidebar() {
   const navigate = useNavigate();
