+++ src/components/pages/Signup.js
@@ -1,0 +1,14 @@
+import React from "react";
+import Template from "./Template";
+
+function Signup () {
+    return (
+        <div>
+            <Template 
+            formtype='signup'
+            />
+        </div>
+    );
+}
+
+export default Signup;
